Change to base64 env var as default solution (even if it's ugly it's convenient...)

+++ swebench_docker/run_docker.py
@@ -1,4 +1,5 @@
 import asyncio
+import base64
 import json
 import tempfile
 import logging
@@ -11,7 +12,7 @@ from swebench_docker.constants import MAP_VERSION_TO_INSTALL
 logger = logging.getLogger(__name__)
 
 
-async def run_docker_evaluation(task_instance: dict, namespace: str, log_dir: str, timeout: int = 900, log_suffix: str = "", verbose: bool = False):
+async def run_docker_evaluation(task_instance: dict, namespace: str, log_dir: str, timeout: int = 900, log_suffix: str = "", verbose: bool = False, base64_instance: bool = True):
     repo_name = task_instance['repo'].replace("/", "_")
 
     specifications = MAP_VERSION_TO_INSTALL[task_instance["repo"]][task_instance["version"]]
@@ -28,31 +29,46 @@ async def run_docker_evaluation(task_instance: dict, namespace: str, log_dir: st
     else:
         docker_image = f"{namespace}/{image_prefix}-{repo_name}-testbed:{task_instance['version']}"
 
-    # Create a temporary file to store the task_instance JSON
-    tmpfile_path = tempfile.mktemp(suffix='.json')
-    with open(tmpfile_path, 'w+') as f:
-        json.dump(task_instance, f)
-
     swebench_docker_fork_dir = os.environ.get("SWEBENCH_DOCKER_FORK_DIR")
-    docker_command = [
-        'docker', 'run',
-        '--rm',
-        '-u', 'root',
-        '-v', f"{log_dir}:{container_log_dir}",
-        # Map the swebench_docker fork dir to the container
-        # for some reason, swebench_docker has different locations for the different containers :(
-        # so we need to map all of them to make it work
-        '-v', f"{swebench_docker_fork_dir}/swebench_docker:/opt/swebench_docker:ro",
-        '-v', f"{swebench_docker_fork_dir}/swebench_docker:/home/swe-bench/swebench_docker:ro",
-        '-v', f"{swebench_docker_fork_dir}/swebench_docker:/home/swe-bench/swebench:ro",
-        # =======
-        # Map file instead pass the instance as env var to avoid "Argument list too long" error
-        '-v', f"{tmpfile_path}:/home/swe-bench/task_instance.json:ro",
-        '-e', f"LOG_DIR={container_log_dir}",
-        '-e', f"TIMEOUT={timeout}",
-        '-e', f"LOG_SUFFIX={log_suffix}",
-        docker_image
-    ]
+
+    if swebench_docker_fork_dir:
+        # Create a temporary file to store the task_instance JSON
+        tmpfile_path = tempfile.mktemp(suffix='.json')
+        with open(tmpfile_path, 'w+') as f:
+            json.dump(task_instance, f)
+
+        docker_command = [
+            'docker', 'run',
+            '--rm',
+            '-u', 'root',
+            '-v', f"{log_dir}:{container_log_dir}",
+            # Map the swebench_docker fork dir to the container
+            # for some reason, swebench_docker has different locations for the different containers :(
+            # so we need to map all of them to make it work
+            '-v', f"{swebench_docker_fork_dir}/swebench_docker:/opt/swebench_docker:ro",
+            '-v', f"{swebench_docker_fork_dir}/swebench_docker:/home/swe-bench/swebench_docker:ro",
+            '-v', f"{swebench_docker_fork_dir}/swebench_docker:/home/swe-bench/swebench:ro",
+            # =======
+            # Map file instead pass the instance as env var to avoid "Argument list too long" error
+            '-v', f"{tmpfile_path}:/home/swe-bench/task_instance.json:ro",
+            '-e', f"LOG_DIR={container_log_dir}",
+            '-e', f"TIMEOUT={timeout}",
+            '-e', f"LOG_SUFFIX={log_suffix}",
+            docker_image
+        ]
+    else:
+        # Base64 encode the instance JSON to be sure it can be passed as an environment variable
+        instance_b64 = base64.b64encode(json.dumps(task_instance).encode('utf-8')).decode('utf-8')
+        docker_command = [
+            'docker', 'run',
+            '--rm',
+            '-v', f"{log_dir}:{container_log_dir}",
+            '-e', f"INSTANCE={instance_b64}",
+            '-e', f"LOG_DIR={container_log_dir}",
+            '-e', f"TIMEOUT={timeout}",
+            '-e', f"LOG_SUFFIX={log_suffix}",
+            docker_image
+        ]
 
     cmd_string = ' '.join(docker_command)
 
@@ -88,5 +104,6 @@ async def run_docker_evaluation(task_instance: dict, namespace: str, log_dir: st
     except Exception as e:
         logger.warning(f"[{task_instance['instance_id']}][{docker_image}]  Error running container: {e}")
     finally:
-        # Ensure the temporary file is deleted after the Docker process completes
-        os.unlink(tmpfile_path)
+        if swebench_docker_fork_dir:
+            # Ensure the temporary file is deleted after the Docker process completes
+            os.unlink(tmpfile_path)
